pyproject.toml
[tool.poetry]
name = "flaskproj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-WTF = "^1.0.1"
email-validator = "^1.2.1"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2 = "^2.9.3"
Flask-Bootstrap = "^3.3.7"
Flask-Login = "^0.6.2"
Flask-Admin = "^1.6.0"
requests = "^2.28.1"
Flask-Mail = "^0.9.1"
python-dotenv = "^0.21.0"
itsdangerous = "^2.1.2"
Flask-BabelEx = "^0.9.4"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
